+++ src/carbon_txt/validators.py
@@ -22,6 +22,19 @@ class ValidationResult:
     exceptions: list
 
 
+def safely_log_exception(exception: Exception, message: str, errors: list, logs: list):  # noqa
+    """
+    Log an an exception, and append it to a list of errors in a form that can
+    be displayed as JSON.
+    """
+    dumpable_error = f"{type(exception).__name__}: {exception}"
+    logger.warning(message)
+    logs.append(message)
+    errors.append(dumpable_error)
+
+    return errors, logs
+
+
 class CarbonTxtValidator:
     """
     The core class repsonsible for exposing essentially the same functionality to
@@ -38,10 +51,11 @@ class CarbonTxtValidator:
         or raises a list of validation exceptions if the contents are invalid.
         """
         self.event_log = []  # Reset event log for each validation
-
         errors: list[Union[Exception, pydantic_core.ErrorDetails, dict]] = []
 
         try:
+            message = f"Attempting to validate contents of {contents[:40]}"
+            self.event_log.append(message)
             parsed_result = parser.parse_toml(contents, logs=self.event_log)
             validation_results = parser.validate_as_carbon_txt(
                 parsed_result, logs=self.event_log
@@ -50,13 +64,16 @@ class CarbonTxtValidator:
                 result=validation_results, logs=self.event_log, exceptions=errors
             )
         except pydantic.ValidationError as ex:
+            message = f"Validation error: {ex}"
+            self.event_log.append(message)
             errors.append(*ex.errors())
             validation_results = None
             return ValidationResult(
                 result=validation_results, logs=self.event_log, exceptions=errors
             )
         except Exception as ex:
-            errors.append(ex)
+            message = f"An unexpected error occurred: {ex}"
+            safely_log_exception(ex, message, errors, self.event_log)
             validation_results = None
             return ValidationResult(
                 result=validation_results, logs=self.event_log, exceptions=errors
@@ -66,9 +83,12 @@ class CarbonTxtValidator:
         """
         Validate a carbon.txt file at a given URL.
         """
+        self.event_log = []  # Reset event log for each validation
         errors: list[Union[Exception, pydantic_core.ErrorDetails, dict]] = []
 
         try:
+            message = f"Attempting to validate url: {url}"
+            self.event_log.append(message)
             result = file_finder.resolve_uri(url, logs=self.event_log)
             fetched_file_contents = file_finder.fetch_carbon_txt_file(
                 result, logs=self.event_log
@@ -87,9 +107,8 @@ class CarbonTxtValidator:
         except FileNotFoundError as ex:
             full_file_path = pathlib.Path(url).absolute()
             message = f"No valid carbon.txt file found at {full_file_path}. \n"
+            safely_log_exception(ex, message, errors, self.event_log)
             validation_results = None
-            self.event_log.append(message)
-            errors.append(ex)
             return ValidationResult(
                 result=validation_results, logs=self.event_log, exceptions=errors
             )
@@ -97,7 +116,7 @@ class CarbonTxtValidator:
         # we have a valid TOML file, but it's not a valid carbon.txt file
         except pydantic.ValidationError as ex:
             message = f"Validation error: {ex}"
-            logger.warning(message)
+            self.event_log.append(message)
             errors.append(*ex.errors())
             validation_results = None
             return ValidationResult(
@@ -106,8 +125,8 @@ class CarbonTxtValidator:
 
         # the file path is remote, and we can't access it
         except exceptions.UnreachableCarbonTxtFile as ex:
-            # logger.error(f"Error: {ex}")
-            errors.append(ex)
+            message = f"Could not fetch the carbon.txt file at {url}. Error was: {ex}"
+            safely_log_exception(ex, message, errors, self.event_log)
             validation_results = None
             return ValidationResult(
                 result=validation_results, logs=self.event_log, exceptions=errors
@@ -117,10 +136,7 @@ class CarbonTxtValidator:
         # with the url listed in the error message, so it's clear to what URL the error refers to
         except exceptions.NotParseableTOML as ex:
             message = f"A file was found at {url}: but it wasn't parseable TOML. Error was: {ex}"
-            logger.warning(message)
-            self.event_log.append(message)
-
-            errors.append(ex)
+            safely_log_exception(ex, message, errors, self.event_log)
             validation_results = None
             return ValidationResult(
                 result=validation_results, logs=self.event_log, exceptions=errors
@@ -129,19 +145,20 @@ class CarbonTxtValidator:
         # the file path is reachable, but the server returned a 404
         except httpx.HTTPStatusError as ex:
             message = f"An error occurred while fetching the carbon.txt file at {url}."
-            logger.warning(message)
-            self.event_log.append(message)
-            errors.append(ex)
+            errors, logs = safely_log_exception(
+                ex, message, errors, logs=self.event_log
+            )
             validation_results = None
             return ValidationResult(
-                result=validation_results, logs=self.event_log, exceptions=errors
+                result=validation_results, logs=logs, exceptions=errors
             )
 
         except Exception as ex:
             message = f"An unexpected error occurred: {ex}"
-            logger.warning(message)
-            self.event_log.append(message)
-            errors.append(ex)
+            errors, logs = safely_log_exception(
+                ex, message, errors, logs=self.event_log
+            )
+            validation_results = None
             return ValidationResult(
                 result=validation_results, logs=self.event_log, exceptions=errors
             )
@@ -156,6 +173,8 @@ class CarbonTxtValidator:
         errors: list[Union[Exception, pydantic_core.ErrorDetails, dict]] = []
 
         try:
+            message = f"Attempting to resolve domain: {domain}"
+            self.event_log.append(message)
             result = file_finder.resolve_domain(domain, logs=self.event_log)
             fetched_file_contents = file_finder.fetch_carbon_txt_file(
                 result, logs=self.event_log
@@ -168,7 +187,8 @@ class CarbonTxtValidator:
                 result=validation_results, logs=self.event_log, exceptions=errors
             )
         except Exception as ex:
-            errors.append(ex)
+            message = f"An unexpected error occurred: {ex}"
+            safely_log_exception(ex, message, errors, self.event_log)
             validation_results = None
             return ValidationResult(
                 result=validation_results, logs=self.event_log, exceptions=errors